Pull fixing methods up into linter
