+++ src/sqlfluff/core/linter/linter.py
@@ -32,6 +32,7 @@ from sqlfluff.core.errors import (
 )
 from sqlfluff.core.file_helpers import get_encoding
 from sqlfluff.core.linter.common import ParsedString, RenderedFile, RuleTuple
+from sqlfluff.core.linter.fix import apply_fixes, compute_anchor_edit_info
 from sqlfluff.core.linter.linted_dir import LintedDir
 from sqlfluff.core.linter.linted_file import (
     TMP_PRS_ERROR_TYPES,
@@ -42,7 +43,6 @@ from sqlfluff.core.linter.linting_result import LintingResult
 from sqlfluff.core.parser import Lexer, Parser
 from sqlfluff.core.parser.segments.base import BaseSegment, SourceFix
 from sqlfluff.core.rules import BaseRule, RulePack, get_ruleset
-from sqlfluff.core.rules.fix import apply_fixes, compute_anchor_edit_info
 from sqlfluff.core.rules.noqa import IgnoreMask
 
 if TYPE_CHECKING:  # pragma: no cover
